fix • a few adjustments

+++ DEMO_gray.m
@@ -6,12 +6,12 @@
 clear; close all; clc;
 UPSAMP.factor = 4;  % magnification factor
 UPSAMP.iter = 3;  % times of iteration
-DECONV.lambda_1 = 0.01;  % lambda_1
-DECONV.lambda_2 = 1;  % lambda_2
-DECONV.iter_max =10;  % max deconv iterations
 GAU.size = 11;  % size of gaussian kernel
-GAU.var = 1;  % variance of gaussian kernel
-IMG.name = 'words';
+GAU.var = 1.5;  % variance of gaussian kernel
+DECONV.lambda_1 = 0.01;  % lambda_1
+DECONV.lambda_2 = 3;  % lambda_2
+DECONV.iter_max =10;  % max deconv iterations
+IMG.name = 'ceremony';
 IMG.read = ['images/', IMG.name, '_small.jpg'];
 
 %% Load original image and split channels
@@ -31,6 +31,7 @@ title('Original image has been resized to display.');
 
 %% Upsample initially
 H_tilde = im2double(imresize(L, UPSAMP.factor, 'bicubic'));
+H_bicubic = H_tilde;
 
 %% Feed-back control loop (**essential step**)
 for i = 1:UPSAMP.iter
@@ -40,7 +41,8 @@ for i = 1:UPSAMP.iter
     %¡¾Non-blind deconvolution¡¿
     H_star = nbDeconv(H_tilde_norm, DECONV, GAU);
     IMG.appro = 'mine';
-    IMG.other = [num2str(DECONV.lambda_2), '_lam2'];
+    IMG.other = [num2str(DECONV.lambda_2), '_lam2',...
+                 '_', num2str(DECONV.iter_max), '_inner'];
 
 %     %¡¾Non-blind deconvolution: Lagrangian approach¡¿
 %     kernel = fspecial('gaussian', GAU.size, GAU.var);
@@ -72,17 +74,40 @@ figure(FIG_HANDLE.H);
 imshow(H_star);
 title(['After ', num2str(i), ' iterations']);
 
-mkdir(['results/', IMG.name]);
-IMG.write = ['results/', IMG.name, '/',...
-    [IMG.name, '_', IMG.appro,],...
+% Write in the disk
+IMG.folder = ['results/', IMG.name, '/'];
+IMG.src = [IMG.folder, IMG.name, '_lr.jpg'];
+IMG.bicubic = [IMG.folder, IMG.name, '_bicubic',...
+    ['_', num2str(UPSAMP.factor), 'x'], '.jpg'];
+IMG.write = [IMG.folder, IMG.name,...
+    ['_', IMG.appro,],...
     ['_', num2str(UPSAMP.factor), 'x'],...
-    ['_', num2str(UPSAMP.iter), '_iter'],...
+    ['_', num2str(UPSAMP.iter), '_outer'],...
     ['_', num2str(GAU.size),'-', num2str(GAU.var), '_gau'],...
     ['_', IMG.other],...
     '.jpg'];
-imwrite(H_star, IMG.write);
-disp([newline,' ============>¡¾Image "',...
-    IMG.write, '" saved!¡¿',newline]);
+% Create folder
+if ~isfolder(IMG.folder)
+    mkdir(IMG.folder)
+end
+% Save origin image
+if ~isfile(IMG.src)
+    imwrite(L_show, IMG.src);
+    disp([newline,' ============>¡¾Image "',...
+        IMG.src, '" saved!¡¿',newline]);
+end
+% Save bicubic image
+if ~isfile(IMG.bicubic)
+    imwrite(H_bicubic, IMG.bicubic);
+    disp([newline,' ============>¡¾Image "',...
+        IMG.bicubic, '" saved!¡¿',newline]);
+end
+% Save result image
+if ~isfile(IMG.write)
+    imwrite(H_star, IMG.write);
+    disp([newline,' ============>¡¾Image "',...
+        IMG.write, '" saved!¡¿',newline]);
+end
 
 %% Function on call
 function [I, low, gap] = simpnormimg(G)
